# openapi_iaservice.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClaimVision IA Service",
    "version": "3.5.0"
  },
  "paths": {
    "/api/v1/predict": {
      "post": {
        "tags": [
          "No Supervised"
        ],
        "summary": "Predecir tipo de daño vehicular",
        "description": "Recibe una imagen de daño vehicular, la procesa a través del encoder y K-Means, y devuelve el tipo de daño, severidad y confianza.",
        "operationId": "predict_api_v1_predict_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_predict_api_v1_predict_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history": {
      "get": {
        "tags": [
          "No Supervised"
        ],
        "summary": "Consultar historial de inferencias",
        "description": "Devuelve una lista paginada de todas las inferencias realizadas.",
        "operationId": "history_api_v1_history_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número de página",
              "default": 1,
              "title": "Page"
            },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Elementos por página",
              "default": 20,
              "title": "Limit"
            },
            "description": "Elementos por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/retrain": {
      "post": {
        "tags": [
          "No Supervised"
        ],
        "summary": "Re-entrenar K-Means",
        "description": "[Admin] Recibe un conjunto de imágenes y re-entrena el modelo K-Means. Requiere al menos K imágenes.",
        "operationId": "retrain_api_v1_retrain_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_retrain_api_v1_retrain_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/predict": {
      "post": {
        "tags": [
          "Supervised"
        ],
        "summary": "[v2] Clasificar daño vehicular",
        "description": "Usa modelo supervisado (ResNet18) para clasificar el daño en 7 clases incluyendo Sin Daño.",
        "operationId": "predict_v2_api_v2_predict_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_predict_v2_api_v2_predict_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/retrain": {
      "post": {
        "tags": [
          "Supervised"
        ],
        "summary": "[v2] Re-entrenar modelo supervisado",
        "description": "Inicia entrenamiento async del clasificador. Enviar labels como JSON + imágenes.",
        "operationId": "retrain_v2_api_v2_retrain_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_retrain_v2_api_v2_retrain_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2RetrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/retrain/{job_id}": {
      "get": {
        "tags": [
          "Supervised"
        ],
        "summary": "[v2] Estado del re-entreno",
        "description": "Consulta el progreso del entrenamiento asíncrono.",
        "operationId": "retrain_status_api_v2_retrain__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2RetrainStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/history": {
      "get": {
        "tags": [
          "Supervised"
        ],
        "summary": "[v2] Historial de inferencias",
        "description": "Devuelve lista paginada de todas las clasificaciones realizadas.",
        "operationId": "history_v2_api_v2_history_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número de página",
              "default": 1,
              "title": "Page"
            },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Elementos por página",
              "default": 20,
              "title": "Limit"
            },
            "description": "Elementos por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_predict_api_v1_predict_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Imagen del daño vehicular (JPG/PNG)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_predict_api_v1_predict_post"
      },
      "Body_predict_v2_api_v2_predict_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Imagen del vehículo (JPG/PNG)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_predict_v2_api_v2_predict_post"
      },
      "Body_retrain_api_v1_retrain_post": {
        "properties": {
          "k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 2.0,
            "title": "K",
            "description": "Número de clústeres"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "Imágenes para re-entrenamiento (mínimo K)"
          }
        },
        "type": "object",
        "required": [
          "k",
          "files"
        ],
        "title": "Body_retrain_api_v1_retrain_post"
      },
      "Body_retrain_v2_api_v2_retrain_post": {
        "properties": {
          "labels": {
            "type": "string",
            "title": "Labels",
            "description": "JSON: [{\"filename\":\"x.jpg\",\"label\":3}, ...]"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "Imágenes del dataset"
          },
          "epochs": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Epochs",
            "description": "Cantidad de épocas",
            "default": 40
          },
          "lr": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Lr",
            "description": "Learning rate",
            "default": 0.001
          }
        },
        "type": "object",
        "required": [
          "labels",
          "files"
        ],
        "title": "Body_retrain_v2_api_v2_retrain_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id"
          },
          "tipo_dano": {
            "type": "string",
            "title": "Tipo Dano"
          },
          "severidad": {
            "type": "string",
            "title": "Severidad"
          },
          "confianza": {
            "type": "number",
            "title": "Confianza"
          },
          "distancia_centroide": {
            "type": "number",
            "title": "Distancia Centroide"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "cluster_id",
          "tipo_dano",
          "severidad",
          "confianza",
          "distancia_centroide",
          "created_at"
        ],
        "title": "HistoryItem"
      },
      "HistoryResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ],
        "title": "HistoryResponse"
      },
...
      "PredictResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "tipo_dano": {
            "type": "string",
            "title": "Tipo Dano"
          },
          "severidad": {
            "type": "string",
            "title": "Severidad"
          },
          "confianza": {
            "type": "number",
            "title": "Confianza"
          },
          "distancia_centroide": {
            "type": "number",
            "title": "Distancia Centroide"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "tipo_dano",
          "severidad",
          "confianza",
          "distancia_centroide",
          "created_at"
        ],
        "title": "PredictResponse"
      },
      "RetrainResponse": {
        "properties": {
          "k": {
            "type": "integer",
            "title": "K"
          },
          "silhouette": {
            "type": "number",
            "title": "Silhouette"
          },
          "davies_bouldin": {
            "type": "number",
            "title": "Davies Bouldin"
          },
          "inertia": {
            "type": "number",
            "title": "Inertia"
          },
          "mapping": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Mapping"
          },
          "trained_at": {
            "type": "string",
            "title": "Trained At"
          }
        },
        "type": "object",
        "required": [
          "k",
          "silhouette",
          "davies_bouldin",
          "inertia",
          "mapping",
          "trained_at"
        ],
        "title": "RetrainResponse"
      },
...
      "V2HistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "class_id": {
            "type": "integer",
            "title": "Class Id"
          },
          "tipo_dano": {
            "type": "string",
            "title": "Tipo Dano"
          },
          "severidad": {
            "type": "string",
            "title": "Severidad"
          },
          "confianza": {
            "type": "number",
            "title": "Confianza"
          },
          "prob_dist": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Prob Dist"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "class_id",
          "tipo_dano",
          "severidad",
          "confianza",
          "prob_dist",
          "created_at"
        ],
        "title": "V2HistoryItem"
      },
      "V2HistoryResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/V2HistoryItem"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ],
        "title": "V2HistoryResponse"
      },
      "V2PredictResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "class_id": {
            "type": "integer",
            "title": "Class Id"
          },
          "tipo_dano": {
            "type": "string",
            "title": "Tipo Dano"
          },
          "severidad": {
            "type": "string",
            "title": "Severidad"
          },
          "confianza": {
            "type": "number",
            "title": "Confianza"
          },
          "prob_dist": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Prob Dist"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "class_id",
          "tipo_dano",
          "severidad",
          "confianza",
          "prob_dist",
          "created_at"
        ],
        "title": "V2PredictResponse"
      },
      "V2RetrainResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "V2RetrainResponse"
      },
      "V2RetrainStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_epochs": {
            "type": "integer",
            "title": "Total Epochs"
          },
          "current_epoch": {
            "type": "integer",
            "title": "Current Epoch"
          },
          "best_accuracy": {
            "type": "number",
            "title": "Best Accuracy"
          },
          "loss_history": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Loss History"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "total_epochs",
          "current_epoch",
          "best_accuracy",
          "loss_history",
          "created_at"
        ],
        "title": "V2RetrainStatusResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}